+++ src/main/java/io/opentracing/contrib/jfrtracer/impl/jfr/JfrEmitter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.opentracing.contrib.jfrtracer.jfr;
+package io.opentracing.contrib.jfrtracer.impl.jfr;
 
 /**
  * Interface for something capable of emitting a JFR event.
