+++ src/main/java/io/opentracing/contrib/jfrtracer/impl/wrapper/DelegatingJfrTracer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.opentracing.contrib.jfrtracer;
+package io.opentracing.contrib.jfrtracer.impl.wrapper;
 
 import java.util.logging.Logger;
