+++ src/main/java/io/opentracing/contrib/jfrtracer/impl/wrapper/ScopeWrapper.java
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.opentracing.contrib.jfrtracer;
+package io.opentracing.contrib.jfrtracer.impl.wrapper;
 
 import io.opentracing.Scope;
 import io.opentracing.Span;
-import io.opentracing.contrib.jfrtracer.jfr.JfrEmitter;
+import io.opentracing.contrib.jfrtracer.impl.jfr.JfrEmitter;
 
 /**
  * Wrapper for {@link Scope}.
