+++ src/main/java11/io/opentracing/contrib/jfrtracer/impl/jfr/JfrSpanEmitterImpl.java
@@ -13,26 +13,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.opentracing.contrib.jfrtracer.jfr;
+package io.opentracing.contrib.jfrtracer.impl.jfr;
 
 import jdk.jfr.Event;
 import jdk.jfr.Label;
 import jdk.jfr.Category;
 import jdk.jfr.Description;
+import jdk.jfr.StackTrace;
+
+import javax.management.DescriptorKey;
 
 import io.opentracing.Span;
-import io.opentracing.contrib.jfrtracer.jfr.AbstractJfrEmitterImpl;
 
 /**
  * This is the JDK 9 or later implementation of the JfrEmitter.
  */
-public class JfrScopeEmitterImpl extends AbstractJfrEmitterImpl {
-	private Jdk9ScopeEvent currentEvent;
+public class JfrSpanEmitterImpl extends AbstractJfrSpanEmitterImpl {
+	private Jdk9SpanEvent currentEvent;
 
-	@Label("Scope Event")
-	@Description("Open tracing event corresponding to an activation scope")
+	@Label("Span Event")
+	@Description("Open tracing event corresponding to a span.")
 	@Category("Open Tracing")
-	private static class Jdk9ScopeEvent extends Event {
+	@StackTrace(false)
+	private static class Jdk9SpanEvent extends Event {
 		@Label("Operation Name")
 		private String operationName;
 
@@ -44,17 +47,51 @@ public class JfrScopeEmitterImpl extends AbstractJfrEmitterImpl {
 
 		@Label("Parent Id")
 		private String parentId;
+
+		@Label("Start Thread")
+		@Description("The thread initiating the span")
+		private Thread startThread;
+
+		@Label("End Thread")
+		@Description("The thread ending the span")
+		private Thread endThread;
 	}
 
-	JfrScopeEmitterImpl(Span span) {
+	private static class EndEventCommand implements Runnable {
+		private final Jdk9SpanEvent event;
+
+		public EndEventCommand(Jdk9SpanEvent event) {
+			this.event = event;
+		}
+
+		@Override
+		public void run() {
+			event.commit();
+		}
+	}
+
+	private static class BeginEventCommand implements Runnable {
+		private final Jdk9SpanEvent event;
+
+		public BeginEventCommand(Jdk9SpanEvent event) {
+			this.event = event;
+		}
+
+		@Override
+		public void run() {
+			event.begin();
+		}
+	}
+
+	JfrSpanEmitterImpl(Span span) {
 		super(span);
 	}
 
 	@Override
 	public void close() {
-		if (currentEvent != null) {
-			currentEvent.end();
-			currentEvent.commit();
+		if (currentEvent != null && currentEvent.isEnabled()) {
+			currentEvent.endThread = Thread.currentThread();
+			EXECUTOR.execute(new EndEventCommand(currentEvent));
 			currentEvent = null;
 		} else {
 			LOGGER.warning("Close without start discovered!");
@@ -63,14 +100,15 @@ public class JfrScopeEmitterImpl extends AbstractJfrEmitterImpl {
 
 	@Override
 	public void start(String operationName) {
-		currentEvent = new Jdk9ScopeEvent();
+		currentEvent = new Jdk9SpanEvent();
 		if (currentEvent.isEnabled()) {
 			currentEvent.operationName = operationName;
 			currentEvent.traceId = span.context().toTraceId();
 			currentEvent.spanId = span.context().toSpanId();
 			// currentEvent.parentId = span.context().toParentId();
+			currentEvent.startThread = Thread.currentThread();
 		}
-		currentEvent.begin();
+		EXECUTOR.execute(new BeginEventCommand(currentEvent));
 	}
 
 	@Override
